+++ src/agents/EventListenerI.java
@@ -9,13 +9,13 @@ import simulator.objects.Auction;
 
 public interface EventListenerI extends Runnable {
 	int getId();
-	void newAction(Auction auction, long time);
-	void priceChangeAction(Auction auction, long time);
-	void lossAction(Auction auction, long time);
-	void winAction(Auction auction, long time);
-	void expiredAction(Auction auction, long time);
-	void soldAction(Auction auction, long time);
-	void endSoonAction(Auction auction, long time);
+	void newAction(Auction auction, int time);
+	void priceChangeAction(Auction auction, int time);
+	void lossAction(Auction auction, int time);
+	void winAction(Auction auction, int time);
+	void expiredAction(Auction auction, int time);
+	void soldAction(Auction auction, int time);
+	void endSoonAction(Auction auction, int time);
 	void gotPaidAction(Collection<Payment> paymentSet);
 	void itemReceivedAction(Set<ItemSold> itemSet);
 }
